+++ src/scripts/interfaces/choice-full.ts
@@ -1,7 +1,7 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { StringUntrusted } from './string-untrusted';
-
-export type CustomProperties = Record<string, any> | string;
+import { Types } from './types';
+// eslint-disable-next-line import/no-cycle
+import { GroupFull } from './group-full';
 
 /*
   A disabled choice appears in the choice dropdown but cannot be selected
@@ -16,11 +16,11 @@ export interface ChoiceFull {
   choiceEl?: HTMLElement;
   labelClass?: Array<string>;
   labelDescription?: string;
-  customProperties?: CustomProperties;
+  customProperties?: Types.CustomProperties;
   disabled: boolean;
   active: boolean;
   elementId?: string;
-  groupId: number;
+  group: GroupFull | null;
   label: StringUntrusted | string;
   placeholder: boolean;
   selected: boolean;