+++ src/scripts/templates.ts
@@ -268,7 +268,6 @@ const templates: TemplatesInterface = {
       label = escapeForTemplate(allowHTML, label);
       label += ` (${groupName})`;
       label = { trusted: label };
-      div.dataset.groupId = `${choice.groupId}`;
     }
 
     let describedBy: HTMLElement = div;
@@ -300,13 +299,16 @@ const templates: TemplatesInterface = {
       addClassesToElement(div, placeholder);
     }
 
-    div.setAttribute('role', choice.groupId ? 'treeitem' : 'option');
+    div.setAttribute('role', choice.group ? 'treeitem' : 'option');
 
     div.dataset.choice = '';
     div.dataset.id = choice.id as unknown as string;
     div.dataset.value = rawValue;
     if (selectText) {
       div.dataset.selectText = selectText;
+    }
+    if (choice.group) {
+      div.dataset.groupId = `${choice.group.id}`;
     }
 
     assignCustomProperties(div, choice, false);