+++ test/scripts/reducers/items.test.ts
@@ -13,7 +13,7 @@ describe('reducers/items', () => {
         value: 'Item one',
         label: 'Item one',
         id: 1234,
-        groupId: 1,
+        group: null,
         score: 0,
         rank: 0,
         customProperties: {
@@ -79,7 +79,7 @@ describe('reducers/items', () => {
       state = [
         {
           id: 1,
-          groupId: 0,
+          group: null,
           score: 0,
           rank: 0,
           value: 'Item one',
@@ -93,7 +93,7 @@ describe('reducers/items', () => {
         },
         {
           id: 2,
-          groupId: 0,
+          group: null,
           score: 0,
           rank: 0,
           value: 'Item one',