+++ test/scripts/store/store.test.ts
@@ -14,7 +14,7 @@ function shimStore() {
 
 describe('reducers/store', () => {
   let instance: Store<Options>;
-  let subscribeStub: sinon.SinonStub<[listener: StoreListener], void>;
+  let subscribeStub: sinon.SinonStub<[listener: StoreListener], Store<Options>>;
   let dispatchStub: sinon.SinonStub<[action: AnyAction], void>;
   let getStateStub: sinon.SinonStub<any[], State>;
   let emptyState: State;
@@ -30,7 +30,7 @@ describe('reducers/store', () => {
       items: [
         {
           id: 1,
-          groupId: 0,
+          group: null,
           value: 'Item one',
           label: 'Item one',
           active: false,
@@ -44,7 +44,7 @@ describe('reducers/store', () => {
         },
         {
           id: 2,
-          groupId: 0,
+          group: null,
           value: 'Item two',
           label: 'Item two',
           active: true,
@@ -58,7 +58,7 @@ describe('reducers/store', () => {
         },
         {
           id: 3,
-          groupId: 0,
+          group: null,
           value: 'Item three',
           label: 'Item three',
           active: true,
@@ -75,7 +75,7 @@ describe('reducers/store', () => {
         {
           id: 1,
           elementId: 'choices-test-1',
-          groupId: 0,
+          group: null,
           value: 'Choice 1',
           label: 'Choice 1',
           disabled: false,
@@ -90,7 +90,7 @@ describe('reducers/store', () => {
         {
           id: 2,
           elementId: 'choices-test-2',
-          groupId: 0,
+          group: null,
           value: 'Choice 2',
           label: 'Choice 2',
           disabled: false,